move generated sources to the bottom
I have experienced issues when the generated sources are not the very last source.
Sometimes, the generated sources need to access the package-lock.json so that must be available.
This is not the case here, because the package-lock is part of the upstream repository, but maybe that'll change in the future.
There seem to be no beneftis for leaving the generate-sources at the top.