caught logical flaw.
changed test to $XDG_SESSION_TYPE = "wayland"
Reason: $WAYLAND_DISPLAY would only be set on wayland desktops. Meaning despite the socket not existing on X11 the variable containing the name would also be empty, so the check would be reduced to the name of the containing directory. This would always evaluate to true, thereby breaking the functionality intended in the first place here.

Why was this overlooked? Because I tested it on a wayland desktop by enabling and disabling the wayland socket instead of logging in and out and changing session types.